refactor: LadderGame 분리
