+++ src/main/java/nextstep/ladder/LadderGame.java
@@ -3,76 +3,35 @@ package nextstep.ladder;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class LadderGame {
-
-    private static final String VERTICAL_LINE = "|";
-    private static final String MOVEABLE_FORMAT = "-----";
-    private static final String LINE_BREAK = "\n";
-    private static final String PRINT_FORMAT = "%6s";
     private static final String PLAYER_NAME_SPLIT_DELIMITER = ",";
 
-    public static void main(String[] args) {
-        String playerNames = inputPlayerNames();
-        Players players = Players.of(splitNames(playerNames));
+    private final Players players;
+    private final Lines lines;
 
-        int height = inputHeight();
+    private LadderGame(final Players players, final Lines lines) {
+        this.players = players;
+        this.lines = lines;
+    }
+
+    public static LadderGame of(String playerNames, int height) {
+        Players players = Players.of(splitNames(playerNames));
         Lines lines = Lines.of(players.countOfPerson(), height);
 
-        printGameResults(players, lines);
+        return new LadderGame(players, lines);
+    }
+
+    public Players getPlayers() {
+        return players;
+    }
+
+    public Lines getLines() {
+        return lines;
     }
 
     private static List<String> splitNames(String playerNames) {
         return Arrays.stream(playerNames.split(PLAYER_NAME_SPLIT_DELIMITER))
                 .collect(Collectors.toList());
-    }
-
-    private static void printGameResults(Players players, Lines lines) {
-        System.out.println("실행결과");
-        printPlayers(players);
-        printLines(lines);
-    }
-
-    private static void printPlayers(Players players) {
-        String playerNames = IntStream.range(0, players.countOfPerson())
-                .mapToObj(players::getPlayer)
-                .map(player -> String.format(PRINT_FORMAT, player.getName()))
-                .collect(Collectors.joining());
-
-        System.out.println(playerNames);
-    }
-
-    private static void printLines(Lines lines) {
-        String linesString = IntStream.range(0, lines.getHeight())
-                .mapToObj(targetHeight -> lineToString(lines.getLine(targetHeight)))
-                .collect(Collectors.joining(LINE_BREAK));
-
-        System.out.println(linesString);
-    }
-
-    private static String lineToString(Line line) {
-        return IntStream.range(0, line.size())
-                .mapToObj(index -> pointToString(line.getPoint(index)))
-                .collect(Collectors.joining());
-    }
-
-    private static String pointToString(Point point) {
-        if (point.canMoveLeft()) {
-            return String.format(PRINT_FORMAT, MOVEABLE_FORMAT + VERTICAL_LINE);
-        }
-
-        return String.format(PRINT_FORMAT, VERTICAL_LINE);
-    }
-
-    private static int inputHeight() {
-        System.out.println("최대 사다리 높이는 몇 개인가요?");
-        return 5; // TODO: input
-    }
-
-    private static String inputPlayerNames() {
-        System.out.println("참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)");
-
-        return "pobi,honux,crong,jk"; // TODO: input
     }
 }
