+++ src/main/java/nextstep/ladder/Players.java
@@ -6,14 +6,12 @@ import java.util.stream.Collectors;
 public class Players {
     private List<Player> players;
 
-    public Players(List<Player> players) {
+    private Players(List<Player> players) {
         this.players = players;
     }
 
     public static Players of(List<String> playerNames) {
-        List<Player> players = playerNames.stream()
-                .map(Player::new)
-                .collect(Collectors.toList());
+        List<Player> players = playerNames.stream().map(Player::new).collect(Collectors.toList());
 
         return new Players(players);
     }
